refactor: align file organization with timesheet pages
